+++ packages/dsw-mailer/CHANGELOG.md
@@ -8,6 +8,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [4.18.0]
+
+### Added
+
+- Added locale support for mails
+- Added properties from tenant config to params
+
 ## [4.17.0]
 
 Released for version consistency with other DSW tools.
@@ -344,3 +351,4 @@ Released for version consistency with other DSW tools.
 [4.15.0]: /../../tree/v4.15.0
 [4.16.0]: /../../tree/v4.16.0
 [4.17.0]: /../../tree/v4.17.0
+[4.18.0]: /../../tree/v4.18.0